上次头像bug、日志路径、代码生成器调整

+++ admin-core/src/main/java/com/lrs/core/admin/mapper/UserRoleMapper.java
@@ -17,5 +17,5 @@ import java.util.List;
  */
 @Component
 public interface UserRoleMapper extends BaseMapper<UserRole> {
-      List<Role> getUserRoles(Integer userId);
+      List<Role> getUserRoles(Long userId);
 }
